+++ CHANGELOG.md
@@ -6,6 +6,11 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 
 ## [Unreleased]
 
+### Added
+- Added `strict` and `assumeObjects` flag support
+
+### Changed
+- Target handlebars.c version to 0.6.0
 
 ## [0.7.4] - 2016-09-21
 
@@ -14,7 +19,6 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 
 ### Fixed
 - Compilation fails on PHP 7.1 ([#35](https://github.com/jbboehr/php-handlebars/pull/35))
-
 
 ## [0.7.3] - 2016-05-05